advances on testing suite structure, patched check_allow_list_contraints.rs bug
