+++ programs/libreplex_editions_controls/src/logic/check_allow_list_constraints.rs
@@ -28,10 +28,15 @@ pub fn check_allow_list_constraints(
             if let (Some(phase_list_price), Some(phase_max_claims)) = (allow_list_price, allow_list_max_claims) {
                 /// 1. check constraints
                 /// dev: notice that if phase_max_claims is 0, this constraint is disabled
+                msg!("check_allow_list_constraints: phase_max_claims: {}", phase_max_claims);
+                msg!("check_allow_list_constraints: mint_count: {}", minter_stats_phase.mint_count);
+
                 if phase_max_claims > 0 && minter_stats_phase.mint_count >= phase_max_claims {
                     return Err(EditionsControlsError::ExceededAllowListMaxClaims.into());
                 }
-                
+
+                msg!("check_allow_list_constraints: passed phase_max_claims");
+
                 /// 2. construct leaf 
                 let leaf = hashv(&[
                     &minter.to_bytes(),
