+++ test/data/merkle_tree.json
@@ -1,186 +1,39 @@
 {
   "merkle_root": [
-    125,
-    184,
-    194,
-    116,
-    52,
-    36,
-    65,
-    219,
-    171,
-    135,
-    154,
-    27,
-    188,
-    122,
-    207,
-    204,
-    111,
-    70,
-    66,
-    115,
-    161,
-    228,
-    44,
-    84,
-    67,
-    97,
-    29,
-    70,
-    253,
-    69,
-    11,
-    245
+    125, 184, 194, 116, 52, 36, 65, 219, 171, 135, 154, 27, 188, 122, 207, 204,
+    111, 70, 66, 115, 161, 228, 44, 84, 67, 97, 29, 70, 253, 69, 11, 245
   ],
   "max_num_nodes": 2,
   "max_total_claim": 6,
   "tree_nodes": [
     {
       "claimant": [
-        104,
-        164,
-        83,
-        51,
-        23,
-        177,
-        193,
-        29,
-        252,
-        241,
-        86,
-        132,
-        173,
-        155,
-        114,
-        131,
-        130,
-        73,
-        27,
-        101,
-        233,
-        95,
-        12,
-        45,
-        107,
-        255,
-        120,
-        26,
-        121,
-        221,
-        120,
-        54
+        104, 164, 83, 51, 23, 177, 193, 29, 252, 241, 86, 132, 173, 155, 114,
+        131, 130, 73, 27, 101, 233, 95, 12, 45, 107, 255, 120, 26, 121, 221,
+        120, 54
       ],
       "claim_price": 500000,
       "max_claims": 3,
       "proof": [
         [
-          64,
-          131,
-          242,
-          169,
-          206,
-          112,
-          155,
-          119,
-          81,
-          214,
-          17,
-          137,
-          174,
-          140,
-          208,
-          220,
-          141,
-          177,
-          213,
-          131,
-          127,
-          104,
-          181,
-          15,
-          121,
-          228,
-          87,
-          25,
-          232,
-          172,
-          235,
-          168
+          64, 131, 242, 169, 206, 112, 155, 119, 81, 214, 17, 137, 174, 140,
+          208, 220, 141, 177, 213, 131, 127, 104, 181, 15, 121, 228, 87, 25,
+          232, 172, 235, 168
         ]
       ]
     },
     {
       "claimant": [
-        186,
-        39,
-        133,
-        92,
-        39,
-        241,
-        42,
-        161,
-        180,
-        15,
-        92,
-        18,
-        15,
-        101,
-        248,
-        80,
-        238,
-        254,
-        220,
-        231,
-        1,
-        14,
-        231,
-        145,
-        170,
-        49,
-        163,
-        111,
-        239,
-        112,
-        135,
-        6
+        186, 39, 133, 92, 39, 241, 42, 161, 180, 15, 92, 18, 15, 101, 248, 80,
+        238, 254, 220, 231, 1, 14, 231, 145, 170, 49, 163, 111, 239, 112, 135, 6
       ],
       "claim_price": 500000,
       "max_claims": 3,
       "proof": [
         [
-          86,
-          37,
-          15,
-          136,
-          192,
-          159,
-          125,
-          244,
-          163,
-          213,
-          251,
-          242,
-          217,
-          215,
-          159,
-          249,
-          93,
-          166,
-          82,
-          38,
-          187,
-          58,
-          199,
-          64,
-          161,
-          50,
-          122,
-          122,
-          17,
-          125,
-          63,
-          188
+          86, 37, 15, 136, 192, 159, 125, 244, 163, 213, 251, 242, 217, 215,
+          159, 249, 93, 166, 82, 38, 187, 58, 199, 64, 161, 50, 122, 122, 17,
+          125, 63, 188
         ]
       ]
     }
