+++ test/utils/getters.ts
@@ -1,9 +1,10 @@
-import { Connection, PublicKey } from '@solana/web3.js';
-import { LibreplexEditions } from '../target/types/libreplex_editions';
-import { IdlAccounts } from '@coral-xyz/anchor';
-import { BorshCoder, Program } from '@coral-xyz/anchor';
-import { LibreplexEditionsControls } from '../target/types/libreplex_editions_controls';
 import { TokenMetadata } from '@solana/spl-token-metadata';
+import { BorshCoder, Program } from '@coral-xyz/anchor';
+import { Connection, PublicKey } from '@solana/web3.js';
+import { LibreplexEditionsControls } from '../../target/types/libreplex_editions_controls';
+import { LibreplexEditions } from '../../target/types/libreplex_editions';
+import { IdlAccounts } from '@coral-xyz/anchor';
+import { getTokenMetadata as getSplTokenMetadata } from '@solana/spl-token';
 
 export type EditionsDeployment =
   IdlAccounts<LibreplexEditions>['editionsDeployment'];
@@ -12,30 +13,6 @@ export type EditionsControls =
   IdlAccounts<LibreplexEditionsControls>['editionsControls'];
 
 export type MinterStats = IdlAccounts<LibreplexEditionsControls>['minterStats'];
-
-export async function getCluster(connection: Connection): Promise<string> {
-  // Get the genesis hash
-  const genesisHash = await connection.getGenesisHash();
-
-  // Compare the genesis hash with known cluster genesis hashes
-  switch (genesisHash) {
-    case '5eykt4UsFv8P8NJdTREpY1vzqKqZKvdpKuc147dw2N9d':
-      return 'mainnet-beta';
-    case 'EtWTRABZaYq6iMfeYKouRu166VU2xqa1wcaWoxPkrZBG':
-      return 'testnet';
-    case '4uhcVJyU9pJkvQyS88uRDiswHXSCkY3zQawwpjk2NsNY':
-      return 'devnet';
-    default:
-      // If it doesn't match any known cluster, it's likely localhost
-      return 'localhost';
-  }
-}
-
-export const getBase64FromDatabytes = (dataBytes: Buffer, dataType: string) => {
-  console.log({ dataBytes });
-  const base = dataBytes.toString('base64');
-  return `data:${dataType};base64,${base}`;
-};
 
 export const decodeEditions =
   (program: Program<LibreplexEditions>) =>
@@ -149,19 +126,6 @@ export const logEditionsControls = (editionsControlsDecoded: {
   });
 };
 
-export const logTokenMetadata = (metadata: TokenMetadata) => {
-  console.log({
-    TokenMetadata: {
-      name: metadata.name,
-      symbol: metadata.symbol,
-      uri: metadata.uri,
-      updateAuthority: metadata.updateAuthority.toBase58(),
-      mint: metadata.mint.toBase58(),
-      additionalMetadata: parseMetadata(metadata.additionalMetadata),
-    },
-  });
-};
-
 export const parseMetadata = (
   rawMetadata: (readonly [string, string])[]
 ): Record<string, string> => {
@@ -170,6 +134,39 @@ export const parseMetadata = (
     metadata[key] = value;
   }
   return metadata;
+};
+
+export const getTokenMetadata = async (
+  connection: Connection,
+  mint: PublicKey
+) => {
+  const rawMetadata = await getSplTokenMetadata(connection, mint);
+  const additionalMetadata = rawMetadata.additionalMetadata;
+  const parsedMetadata = parseMetadata(additionalMetadata);
+  return {
+    ...rawMetadata,
+    additionalMetadata: parsedMetadata,
+  };
+};
+
+export const logTokenMetadata = (metadata: {
+  name: string;
+  symbol: string;
+  uri: string;
+  updateAuthority?: PublicKey;
+  mint: PublicKey;
+  additionalMetadata: Record<string, string>;
+}) => {
+  console.log({
+    TokenMetadata: {
+      name: metadata.name,
+      symbol: metadata.symbol,
+      uri: metadata.uri,
+      updateAuthority: metadata.updateAuthority.toBase58(),
+      mint: metadata.mint.toBase58(),
+      additionalMetadata: metadata.additionalMetadata,
+    },
+  });
 };
 
 export const decodeMinterStats =
